Refactor ui
